Finally got seperate thread for bot working properly! In addition, timed moves now implemented: bot will take a maximum of 10 seconds to make its move (by default).
I have found a bug in the bot's static position evaluator, an
ArrayOutOfBounds can happen in ROW_SCORES and I'm not sure why. It's not
lethal to the program because the botMove wrapper handles execution
exceptions. It does mean that the bot will occasionally make stupid
moves until I can find and fix the bug.